# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "simple intranet dns management system, dns crud operation with centos bind9",
        "title": "sre dns service backend demo",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:16789",
    "paths": {
...
        "/api/v1/cronjobs": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cronjob"
                ],
                "summary": "list cronjob",
                "parameters": [
                    {
                        "type": "string",
                        "description": "min=1",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "page, min=1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size, min=10, max=1000",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cronjob name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cronjob creator",
                        "name": "creator",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "cronjob type",
                        "name": "task_type",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "cronjob started",
                        "name": "started",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "cronjob last running status",
                        "name": "last_succeed",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "users",
                        "schema": {
                            "$ref": "#/definitions/ctx.StdResponse"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cronjob"
                ],
                "summary": "update cronjob",
                "parameters": [
                    {
                        "type": "string",
                        "description": "min=1",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "description": "update cronjob request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.updateCronjobReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "cronjob id",
                        "schema": {
                            "$ref": "#/definitions/ctx.StdResponse"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cronjob"
                ],
                "summary": "add cronjob",
                "parameters": [
                    {
                        "type": "string",
                        "description": "min=1",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "description": "cronjob request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apis.newCronjob"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "cronjob id",
                        "schema": {
                            "$ref": "#/definitions/ctx.StdResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/cronjobs/{id}": {
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cronjob"
                ],
                "summary": "del cronjob",
                "parameters": [
                    {
                        "type": "string",
                        "description": "min=1",
                        "name": "token",
                        "in": "header"
                    },
                    {
                        "type": "integer",
                        "description": "cronjob id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "cronjob id",
                        "schema": {
                            "$ref": "#/definitions/ctx.StdResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apis.newCronjob": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "task desc",
                    "type": "string"
                },
                "name": {
                    "description": "task name",
                    "type": "string"
                },
                "spec": {
                    "description": "spec \u0026 s m h D M W",
                    "type": "string",
                    "minLength": 9
                },
                "taskArgs": {
                    "description": "task args",
                    "$ref": "#/definitions/models.Args"
                },
                "taskType": {
                    "description": "task type",
                    "type": "string"
                }
            }
        },
...
        "apis.updateCronjobReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "description": {
                    "description": "task desc",
                    "type": "string"
                },
                "id": {
                    "description": "user id",
                    "type": "integer"
                },
                "name": {
                    "description": "task name",
                    "type": "string"
                },
                "spec": {
                    "description": "spec \u0026 s m h D M W",
                    "type": "string"
                },
                "started": {
                    "description": "task switch",
                    "type": "boolean"
                },
                "task_args": {
                    "description": "task args",
                    "$ref": "#/definitions/models.Args"
                }
            }
        },
...
        "ctx.StdResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "data"
                },
                "error": {
                    "description": "self err",
                    "type": "string"
                },
                "pages": {
                    "description": "pages",
                    "$ref": "#/definitions/models.PageRsp"
                },
                "request_id": {
                    "description": "api request uid",
                    "type": "string"
                },
                "status": {
                    "description": "true: succeed, false: failed",
                    "type": "boolean"
                }
            }
        },
...
        "models.Args": {
            "type": "object",
            "properties": {
                "function_name": {
                    "description": "function name",
                    "type": "string"
                },
                "url": {
                    "description": "http url",
                    "type": "string"
                }
            }
        },
        "models.PageRsp": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    }
}